+++ api-spec/protobuf/wallet.proto
@@ -6,37 +6,6 @@ option go_package = "github.com/tdex-network/tdex-daemon/api-spec/protobuf/gen/w
  */
 service Wallet {
   /*
-  GenSeed is the first method that should be used to instantiate a new tdexd
-  instance. This method allows a caller to generate a new HD Wallet.
-  Once the seed is obtained and verified by the user, the InitWallet
-  method should be used to commit the newly generated seed, and create the
-  wallet.
-  */
-  rpc GenSeed(GenSeedRequest) returns (GenSeedReply);
-
-  /*
-  InitWallet is used when tdexd is starting up for the first time to fully
-  initialize the daemon and its internal wallet. At the very least a mnemonic
-  and a wallet password must be provided. This will be used to encrypt sensitive
-  material on disk. Alternatively, this can be used along with the GenSeed RPC
-  to obtain a seed, then present it to the user. Once it has been verified by
-  the user, the seed can be fed into this RPC in order to commit the new wallet.
-  */
-  rpc InitWallet(InitWalletRequest) returns (stream InitWalletReply);
-
-  /*
-  UnlockWallet is used at startup of tdexd to provide a password to unlock
-  the wallet database.
-  */
-  rpc UnlockWallet(UnlockWalletRequest) returns (UnlockWalletReply);
-
-  /*
-  ChangePassword changes the password of the encrypted wallet. This will
-  automatically unlock the wallet database if successful.
-  */
-  rpc ChangePassword(ChangePasswordRequest) returns (ChangePasswordReply);
-
-  /*
   WalletAddress returns a Liquid confidential p2wpkh address (BLECH32)
   */
   rpc WalletAddress(WalletAddressRequest) returns (WalletAddressReply);
@@ -51,65 +20,6 @@ service Wallet {
   /*SendToMany sends funds to many outputs */
   rpc SendToMany(SendToManyRequest) returns (SendToManyReply);
 }
-
-message GenSeedRequest {}
-message GenSeedReply {
-  repeated string seed_mnemonic = 1;
-}
-
-message InitWalletRequest {
-  /*
-  wallet_password is the passphrase that should be used to encrypt the
-  wallet. This MUST be at least 8 chars in length. After creation, this
-  password is required to unlock the daemon.
-  */
-  bytes wallet_password = 1;
-
-  /*
-  seed_mnemonic is a 24-word mnemonic that encodes a prior seed obtained by the
-  user. This MUST be a generated by the GenSeed method
-  */
-  repeated string seed_mnemonic = 2;
-  /*
-  the flag to let the daemon restore existing funds for the wallet.
-  */
-  bool restore = 3;
-}
-message InitWalletReply {
-  uint64 account = 1;
-  uint64 index = 2;
-  enum Status {
-    PROCESSING = 0;
-    DONE = 1;
-  } 
-  Status status = 3;
-  string data = 4;
-}
-
-message UnlockWalletRequest {
-  /*
-  wallet_password should be the current valid passphrase for the daemon. This
-  will be required to decrypt on-disk material that the daemon requires to
-  function properly.
-  */
-  bytes wallet_password = 1;
-}
-message UnlockWalletReply {}
-
-message ChangePasswordRequest {
-  /*
-  current_password should be the current valid passphrase used to unlock the
-  daemon.
-  */
-  bytes current_password = 1;
-
-  /*
-  new_password should be the new passphrase that will be needed to unlock the
-  daemon.
-  */
-  bytes new_password = 2;
-}
-message ChangePasswordReply {}
 
 message SendToManyRequest {
   /*
