+++ api-spec/protobuf/walletunlocker.proto
@@ -1,0 +1,119 @@
+syntax = "proto3";
+option go_package = "github.com/tdex-network/tdex-daemon/api-spec/protobuf/gen/walletunlocker";
+
+/**
+ * Service for operators to manage the internal wallet of the daemon
+ */
+service WalletUnlocker {
+  /*
+  GenSeed is the first method that should be used to instantiate a new tdexd
+  instance. This method allows a caller to generate a new HD Wallet.
+  Once the seed is obtained and verified by the user, the InitWallet
+  method should be used to commit the newly generated seed, and create the
+  wallet.
+  */
+  rpc GenSeed(GenSeedRequest) returns (GenSeedReply);
+
+  /*
+  InitWallet is used when tdexd is starting up for the first time to fully
+  initialize the daemon and its internal wallet. 
+  The wallet in the tdexd context is a database file on the disk that can be
+  found in the configured data directory. 
+  At the very least a mnemonic and a wallet password must be provided to this 
+  RPC. The latter will be used to encrypt sensitive material on disk.
+  Once initialized the wallet is locked and since the password is never stored
+  on the disk, it's required to pass it into the Unlock RPC request to be able
+  to manage the daemon for operations like depositing funds or opening a market.
+  */
+  rpc InitWallet(InitWalletRequest) returns (stream InitWalletReply);
+
+  /*
+  UnlockWallet is used at startup of tdexd to provide a password to unlock
+  the wallet database. Once unlocked, the only way to lock the wallet again is
+  shutting it down.
+  */
+  rpc UnlockWallet(UnlockWalletRequest) returns (UnlockWalletReply);
+
+  /*
+  ChangePassword changes the password of the encrypted wallet. This RPC
+  requires the internal wallet to be locked. It doesn't change the wallet state
+  in any case, therefore, like after calling InitWallet, it is required to 
+  unlock the walket with UnlockWallet RPC after this operation succeeds.
+  */
+  rpc ChangePassword(ChangePasswordRequest) returns (ChangePasswordReply);
+
+  /*
+    IsReady is useful for external applications interacting with tdexd to know
+    whether its ready, meaning that also the wallet, operator trade services
+    are able to serve requests.
+    Restarting tdexd or initiliazing it by restoring an existing wallet can be
+    time-expensive operations causing tdexd to not be ready until they haven't
+    finished.
+  */
+  rpc IsReady(IsReadyRequest) returns (IsReadyReply);
+}
+
+
+message GenSeedRequest {}
+message GenSeedReply {
+  repeated string seed_mnemonic = 1;
+}
+
+message InitWalletRequest {
+  /*
+  wallet_password is the passphrase that should be used to encrypt the
+  wallet. This MUST be at least 8 chars in length. After creation, this
+  password is required to unlock the daemon.
+  */
+  bytes wallet_password = 1;
+
+  /*
+  seed_mnemonic is a 24-word mnemonic that encodes a prior seed obtained by the
+  user. This MUST be a generated by the GenSeed method
+  */
+  repeated string seed_mnemonic = 2;
+  /*
+  the flag to let the daemon restore existing funds for the wallet.
+  */
+  bool restore = 3;
+}
+message InitWalletReply {
+  uint64 account = 1;
+  uint64 index = 2;
+  enum Status {
+    PROCESSING = 0;
+    DONE = 1;
+  } 
+  Status status = 3;
+  string data = 4;
+}
+
+message UnlockWalletRequest {
+  /*
+  wallet_password should be the current valid passphrase for the daemon. This
+  will be required to decrypt on-disk material that the daemon requires to
+  function properly.
+  */
+  bytes wallet_password = 1;
+}
+message UnlockWalletReply {}
+
+message ChangePasswordRequest {
+  /*
+  current_password should be the current valid passphrase used to unlock the
+  daemon.
+  */
+  bytes current_password = 1;
+
+  /*
+  new_password should be the new passphrase that will be needed to unlock the
+  daemon.
+  */
+  bytes new_password = 2;
+}
+message ChangePasswordReply {}
+
+message IsReadyRequest {}
+message IsReadyReply {
+  bool is_ready = 1;
+}
