+++ cmd/tdex/genseed.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"strings"
 
-	pbwallet "github.com/tdex-network/tdex-daemon/api-spec/protobuf/gen/wallet"
+	pbwallet "github.com/tdex-network/tdex-daemon/api-spec/protobuf/gen/walletunlocker"
 
 	"github.com/urfave/cli/v2"
 )
