+++ cmd/tdex/initwallet.go
@@ -6,7 +6,7 @@ import (
 	"io"
 	"strings"
 
-	pbwallet "github.com/tdex-network/tdex-daemon/api-spec/protobuf/gen/wallet"
+	pbwallet "github.com/tdex-network/tdex-daemon/api-spec/protobuf/gen/walletunlocker"
 
 	"github.com/urfave/cli/v2"
 )
