+++ internal/interfaces/grpc/handler/wallet.go
@@ -31,34 +31,6 @@ func newWalletHandler(
 	}
 }
 
-func (w walletHandler) GenSeed(
-	ctx context.Context,
-	req *pb.GenSeedRequest,
-) (*pb.GenSeedReply, error) {
-	return w.genSeed(ctx, req)
-}
-
-func (w walletHandler) InitWallet(
-	req *pb.InitWalletRequest,
-	stream pb.Wallet_InitWalletServer,
-) error {
-	return w.initWallet(req, stream)
-}
-
-func (w walletHandler) UnlockWallet(
-	ctx context.Context,
-	req *pb.UnlockWalletRequest,
-) (*pb.UnlockWalletReply, error) {
-	return w.unlockWallet(ctx, req)
-}
-
-func (w walletHandler) ChangePassword(
-	ctx context.Context,
-	req *pb.ChangePasswordRequest,
-) (*pb.ChangePasswordReply, error) {
-	return w.changePassword(ctx, req)
-}
-
 func (w walletHandler) WalletAddress(
 	ctx context.Context,
 	req *pb.WalletAddressRequest,
@@ -78,102 +50,6 @@ func (w walletHandler) SendToMany(
 	req *pb.SendToManyRequest,
 ) (*pb.SendToManyReply, error) {
 	return w.sendToMany(ctx, req)
-}
-
-func (w walletHandler) genSeed(
-	ctx context.Context,
-	req *pb.GenSeedRequest,
-) (*pb.GenSeedReply, error) {
-	mnemonic, err := w.walletSvc.GenSeed(ctx)
-	if err != nil {
-		return nil, status.Error(codes.Internal, err.Error())
-	}
-	return &pb.GenSeedReply{SeedMnemonic: mnemonic}, nil
-}
-
-func (w walletHandler) initWallet(
-	req *pb.InitWalletRequest,
-	stream pb.Wallet_InitWalletServer,
-) error {
-	mnemonic := req.GetSeedMnemonic()
-	if err := validateMnemonic(mnemonic); err != nil {
-		return status.Error(codes.InvalidArgument, err.Error())
-	}
-	password := req.GetWalletPassword()
-	if err := validatePassword(password); err != nil {
-		return status.Error(codes.InvalidArgument, err.Error())
-	}
-
-	chReplies := make(chan *application.InitWalletReply)
-	chErr := make(chan error, 1)
-	go w.walletSvc.InitWallet(
-		stream.Context(),
-		mnemonic,
-		string(password),
-		req.GetRestore(),
-		chReplies,
-		chErr,
-	)
-
-	for {
-		select {
-		case err := <-chErr:
-			return err
-		case reply, ok := <-chReplies:
-			if !ok {
-				return nil
-			}
-			if err := stream.Send(&pb.InitWalletReply{
-				Account: uint64(reply.AccountIndex),
-				Index:   uint64(reply.AddressIndex),
-				Status:  pb.InitWalletReply_Status(reply.Status),
-				Data:    reply.Data,
-			}); err != nil {
-				return err
-			}
-		}
-	}
-}
-
-func (w walletHandler) unlockWallet(
-	ctx context.Context,
-	req *pb.UnlockWalletRequest,
-) (*pb.UnlockWalletReply, error) {
-	password := req.GetWalletPassword()
-	if err := validatePassword(password); err != nil {
-		return nil, status.Error(codes.InvalidArgument, err.Error())
-	}
-
-	if err := w.walletSvc.UnlockWallet(ctx, string(password)); err != nil {
-		return nil, err
-	}
-
-	return &pb.UnlockWalletReply{}, nil
-}
-
-func (w walletHandler) changePassword(
-	ctx context.Context,
-	req *pb.ChangePasswordRequest,
-) (*pb.ChangePasswordReply, error) {
-	currentPwd := req.GetCurrentPassword()
-	if err := validatePassword(currentPwd); err != nil {
-		return nil, status.Error(codes.InvalidArgument, err.Error())
-	}
-	newPwd := req.GetNewPassword()
-	if err := validatePassword(newPwd); err != nil {
-		return nil, status.Error(codes.InvalidArgument, err.Error())
-	}
-
-	if err := w.walletSvc.ChangePassword(
-		ctx,
-		string(currentPwd),
-		string(newPwd),
-	); err != nil {
-		return nil, err
-	}
-
-	return &pb.ChangePasswordReply{}, nil
-
 }
 
 func (w walletHandler) walletAddress(
@@ -246,20 +122,6 @@ func (w walletHandler) sendToMany(
 	}
 
 	return &pb.SendToManyReply{RawTx: rawTx}, nil
-}
-
-func validateMnemonic(mnemonic []string) error {
-	if len(mnemonic) <= 0 {
-		return errors.New("mnemonic is null")
-	}
-	return nil
-}
-
-func validatePassword(password []byte) error {
-	if len(password) <= 0 {
-		return errors.New("password is null")
-	}
-	return nil
 }
 
 func validateOutputs(outputs []*pb.TxOut) error {
