+++ scripts/compile_proto
@@ -15,6 +15,7 @@ rm -rf $DST_DIR
 
 mkdir -p $DST_DIR/operator
 mkdir -p $DST_DIR/wallet
+mkdir -p $DST_DIR/walletunlocker
 mkdir -p $DST_DIR/types
 
 tdex_proto_path=$(go list -f '{{ .Dir }}' -m github.com/tdex-network/tdex-protobuf)
